Move meta.Size to its own file, fix Size.String()
