+++ pkg/apis/meta/v1alpha1/meta.go
@@ -1,10 +1,7 @@
 package v1alpha1
 
 import (
-	"encoding/json"
 	"fmt"
-
-	"github.com/c2h5oh/datasize"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -86,86 +83,6 @@ type UID string
 // String returns the UID in string representation
 func (u UID) String() string {
 	return string(u)
-}
-
-// Size specifies a common unit for data sizes
-type Size struct {
-	datasize.ByteSize
-}
-
-var EmptySize = NewSizeFromBytes(0)
-
-var _ json.Marshaler = &Size{}
-var _ json.Unmarshaler = &Size{}
-
-func NewSizeFromString(str string) (Size, error) {
-	s := Size{}
-	err := s.UnmarshalText([]byte(str))
-	return s, err
-}
-
-func NewSizeFromBytes(bytes uint64) Size {
-	return Size{
-		datasize.ByteSize(bytes),
-	}
-}
-
-func NewSizeFromSectors(sectors uint64) Size {
-	return Size{
-		datasize.ByteSize(sectors * sectorSize),
-	}
-}
-
-func (s *Size) Sectors() uint64 {
-	return s.Bytes() / sectorSize
-}
-
-// Override ByteSize's default string implementation which results in something similar to HR()
-func (s *Size) String() string {
-	b, _ := s.MarshalText()
-	return string(b)
-}
-
-// Int64 returns the byte size as int64
-func (s *Size) Int64() int64 {
-	return int64(s.Bytes())
-}
-
-// Add returns a copy, does not modify the receiver
-func (s Size) Add(other Size) Size {
-	s.ByteSize += other.ByteSize
-	return s
-}
-
-func (s Size) Min(other Size) Size {
-	if other.ByteSize < s.ByteSize {
-		return other
-	}
-
-	return s
-}
-
-func (s Size) Max(other Size) Size {
-	if other.ByteSize > s.ByteSize {
-		return other
-	}
-
-	return s
-}
-
-func (s *Size) MarshalJSON() ([]byte, error) {
-	return json.Marshal(s.String())
-}
-
-func (s *Size) UnmarshalJSON(b []byte) error {
-	var str string
-	err := json.Unmarshal(b, &str)
-	if err != nil {
-		return err
-	}
-
-	*s, err = NewSizeFromString(str)
-	return err
 }
 
 // DMID specifies the format for device mapper IDs
